mark tailwindcss as a peerDependency
Internally we are using tailwindcss/plugin, but we don't mark it as a
peerDependency or a normal dependency in our package.json. This means
that this plugin will fail in a monorepo environment when dependnecies
get hoisted. This is because now there is no link to tailwind itself.

Marking it as a peerDependency should fix this, because now there is a
link. People that already have tailwind installed are good to go. People
that use monorepo's should benefit from this additional information.

Side note: NPM 7 starts installing peerDependencies for us https://github.com/npm/rfcs/blob/latest/implemented/0025-install-peer-deps.md
which is nice, because if we added an `explicit` dependency to Tailwind,
then we would always reference our dependency and not the user's
dependency.

Closes: #31