docs: update installation guide to add tailwind css v4 instructions while keeping v3 details (#180)
This PR updates the installation guide for this plugin,
[@tailwindcss/forms](https://github.com/tailwindlabs/tailwindcss-forms),
to include setup instructions for Tailwind CSS v4 using `@plugin` in the
`style.css` file, while retaining the existing instructions for v3 users
who are yet to upgrade.

**Why this PR?**
Based on my experience, I've had to refer to the typography plugin docs
for setup instructions and adapt them for the forms plugin. This update
ensures the guide is clear and useful for both Tailwind CSS v3 and v4
users.

---------

Co-authored-by: Jordan Pittman <jordan@cryptica.me>